Move user token creation to repository.
